refactor: split utilities in different files
