+++ api/bootstrapper.go
@@ -16,12 +16,9 @@ import (
 	cnpgv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/websocket"
-	"github.com/mattermost/mattermost-cloudnative-bootstrapper/helm"
 	"github.com/mattermost/mattermost-cloudnative-bootstrapper/internal/logger"
 	"github.com/mattermost/mattermost-cloudnative-bootstrapper/model"
-	helmclient "github.com/mittwald/go-helm-client"
 	"helm.sh/helm/v3/pkg/release"
-	"helm.sh/helm/v3/pkg/repo"
 	v1 "k8s.io/api/core/v1"
 	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -371,242 +368,6 @@ func handleGetInstalledCharts(c *Context, w http.ResponseWriter, r *http.Request
 	}
 
 	json.NewEncoder(w).Encode(releasesRes)
-}
-
-func handleDeleteNginxOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "delete-ingress-nginx")
-	c.Ctx = logger.WithNamespace(c.Ctx, "ingress-nginx")
-
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "ingress-nginx")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	err = helmClient.UninstallReleaseByName("ingress-nginx")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to delete ingress-nginx operator")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-
-}
-
-func handleDeployNginxOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "deploy-ingress-nginx")
-	c.Ctx = logger.WithNamespace(c.Ctx, "ingress-nginx")
-
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		model.WriteBadRequestError(w, "Cluster name is required for Nginx operator deployment", "deploy-nginx-operator", map[string]interface{}{
-			"provided_cluster_name": clusterName,
-		})
-		return
-	}
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	// TODO: Before this can run, the subnets that the cluster was created on must be updated to have tags with the format:
-	// kubernetes.io/cluster/cluster-name: shared (TODO: Confirm "shared" is correct?)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "ingress-nginx")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		model.WriteDeploymentError(w, "Failed to connect to the Kubernetes cluster for Nginx operator deployment. Please verify your cluster is running and accessible.", "deploy-nginx-operator", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "ingress-nginx",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	chartRepo := repo.Entry{
-		Name: "nginx",
-		URL:  "https://kubernetes.github.io/ingress-nginx",
-	}
-
-	err = helmClient.AddOrUpdateChartRepo(chartRepo)
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to add or update chart repo")
-		model.WriteDeploymentError(w, "Failed to add the Nginx Helm repository. This may be due to network connectivity issues.", "deploy-nginx-operator", map[string]interface{}{
-			"repository_name": chartRepo.Name,
-			"repository_url":  chartRepo.URL,
-			"cluster_name":    clusterName,
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	// TODO - we need some sort of pre-post hooks for the install to allow for environment specific configurations
-	valuesYaml := helm.NginxOperatorValues
-
-	chartSpec := helmclient.ChartSpec{
-		ReleaseName:     "ingress-nginx",
-		ChartName:       "nginx/ingress-nginx",
-		Namespace:       "ingress-nginx",
-		UpgradeCRDs:     true,
-		Wait:            true,
-		Timeout:         3000 * time.Second,
-		CreateNamespace: true,
-		CleanupOnFail:   true,
-		ValuesYaml:      valuesYaml,
-	}
-
-	// Install a chart release.
-	// Note that helmclient.Options.Namespace should ideally match the namespace in chartSpec.Namespace.
-	if _, err := helmClient.InstallOrUpgradeChart(context.Background(), &chartSpec, nil); err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to install nginx operator")
-		model.WriteDeploymentError(w, "Failed to deploy the Nginx ingress controller to your cluster. This could be due to insufficient resources, networking issues, or missing subnet tags.", "deploy-nginx-operator", map[string]interface{}{
-			"chart_name":      chartSpec.ChartName,
-			"release_name":    chartSpec.ReleaseName,
-			"namespace":       chartSpec.Namespace,
-			"cluster_name":    clusterName,
-			"timeout_seconds": int(chartSpec.Timeout.Seconds()),
-			"error":           err.Error(),
-			"note":            "Ensure your cluster subnets have the kubernetes.io/cluster/cluster-name tag",
-		})
-		return
-	}
-
-	w.WriteHeader(http.StatusCreated)
-
-}
-
-func handleDeletePGOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "delete-cnpg")
-	c.Ctx = logger.WithNamespace(c.Ctx, "cnpg-system")
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "cnpg-system")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	err = helmClient.UninstallReleaseByName("cnpg-system")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to delete cnpg-system operator")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func handleDeployPGOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "deploy-cnpg")
-	c.Ctx = logger.WithNamespace(c.Ctx, "cnpg")
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		model.WriteBadRequestError(w, "Cluster name is required for CloudNative PostgreSQL operator deployment", "deploy-pg-operator", map[string]interface{}{
-			"provided_cluster_name": clusterName,
-		})
-		return
-	}
-
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "kube-system")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		model.WriteDeploymentError(w, "Failed to connect to the Kubernetes cluster for PostgreSQL operator deployment. Please verify your cluster is running and accessible.", "deploy-pg-operator", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "kube-system",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	err = c.CloudProvider.HelmFileStorePre(c.Ctx, clusterName, "kube-system")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to execute file system preinstall steps for cnpg operator")
-		model.WriteDeploymentError(w, "Failed to configure file storage prerequisites for PostgreSQL operator deployment. This may be due to insufficient permissions or storage configuration issues.", "deploy-pg-operator", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "kube-system",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	chartRepo := repo.Entry{
-		Name: "cnpg",
-		URL:  "https://cloudnative-pg.github.io/charts",
-	}
-
-	err = helmClient.AddOrUpdateChartRepo(chartRepo)
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to add or update chart repo")
-		model.WriteDeploymentError(w, "Failed to add the CloudNative PostgreSQL Helm repository. This may be due to network connectivity issues.", "deploy-pg-operator", map[string]interface{}{
-			"repository_name": chartRepo.Name,
-			"repository_url":  chartRepo.URL,
-			"cluster_name":    clusterName,
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	helmClient, err = c.CloudProvider.HelmClient(c.Ctx, clusterName, "cnpg-system")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		model.WriteDeploymentError(w, "Failed to connect to the cnpg-system namespace for PostgreSQL operator deployment.", "deploy-pg-operator", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "cnpg-system",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	chartSpec := helmclient.ChartSpec{
-		ReleaseName:     "cnpg-system",
-		ChartName:       "cnpg/cloudnative-pg",
-		Namespace:       "cnpg-system",
-		UpgradeCRDs:     true,
-		Wait:            true,
-		Timeout:         300 * time.Second,
-		CreateNamespace: true,
-		CleanupOnFail:   true,
-	}
-
-	// Install a chart release.
-	// Note that helmclient.Options.Namespace should ideally match the namespace in chartSpec.Namespace.
-	if _, err := helmClient.InstallOrUpgradeChart(context.Background(), &chartSpec, nil); err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to install cnpg operator")
-		model.WriteDeploymentError(w, "Failed to deploy the CloudNative PostgreSQL operator to your cluster. This could be due to insufficient resources, networking issues, or configuration problems.", "deploy-pg-operator", map[string]interface{}{
-			"chart_name":      chartSpec.ChartName,
-			"release_name":    chartSpec.ReleaseName,
-			"namespace":       chartSpec.Namespace,
-			"cluster_name":    clusterName,
-			"timeout_seconds": int(chartSpec.Timeout.Seconds()),
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	w.WriteHeader(http.StatusCreated)
 }
 
 func handleGetClusterNamespaces(c *Context, w http.ResponseWriter, r *http.Request) {
@@ -1251,209 +1012,6 @@ func handleCreateMattermostInstallation(c *Context, w http.ResponseWriter, r *ht
 		return
 	}
 	json.NewEncoder(w).Encode(mattermost)
-}
-
-func handleDeleteMattermostOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "delete-mattermost")
-	c.Ctx = logger.WithNamespace(c.Ctx, "mattermost-operator")
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "mattermost-operator")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	err = helmClient.UninstallReleaseByName("mattermost-operator")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to delete mattermost-operator operator")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-}
-
-func handleDeployMattermostOperator(c *Context, w http.ResponseWriter, r *http.Request) {
-	logger.FromContext(c.Ctx).Info("Getting kubeconfig")
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-
-	if clusterName == "" {
-		model.WriteBadRequestError(w, "Cluster name is required for Mattermost operator deployment", "deploy-mattermost-operator", map[string]interface{}{
-			"provided_cluster_name": clusterName,
-		})
-		return
-	}
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "mattermost-operator")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		model.WriteDeploymentError(w, "Failed to connect to the Kubernetes cluster for Mattermost operator deployment. Please verify your cluster is running and accessible.", "deploy-mattermost-operator", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "mattermost-operator",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	chartRepo := repo.Entry{Name: "mattermost", URL: "https://helm.mattermost.com"}
-
-	err = helmClient.AddOrUpdateChartRepo(chartRepo)
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to add or update chart repo")
-		model.WriteDeploymentError(w, "Failed to add the Mattermost Helm repository for operator deployment. This may be due to network connectivity issues.", "deploy-mattermost-operator", map[string]interface{}{
-			"repository_name": chartRepo.Name,
-			"repository_url":  chartRepo.URL,
-			"cluster_name":    clusterName,
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	chartSpec := helmclient.ChartSpec{
-		ReleaseName: "mattermost-operator",
-		ChartName:   "mattermost/mattermost-operator",
-		Namespace:   "mattermost-operator",
-		UpgradeCRDs: true,
-		// Version:         "1.25.2",
-		Wait:            true,
-		Timeout:         300 * time.Second,
-		CreateNamespace: true,
-		CleanupOnFail:   true,
-	}
-
-	// Install a chart release.
-	// Note that helmclient.Options.Namespace should ideally match the namespace in chartSpec.Namespace.
-	if _, err := helmClient.InstallOrUpgradeChart(context.Background(), &chartSpec, nil); err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to install mattermost operator")
-		model.WriteDeploymentError(w, "Failed to deploy the Mattermost operator to your cluster. This could be due to insufficient resources, networking issues, or configuration problems.", "deploy-mattermost-operator", map[string]interface{}{
-			"chart_name":      chartSpec.ChartName,
-			"release_name":    chartSpec.ReleaseName,
-			"namespace":       chartSpec.Namespace,
-			"cluster_name":    clusterName,
-			"timeout_seconds": int(chartSpec.Timeout.Seconds()),
-			"error":           err.Error(),
-		})
-		return
-	}
-	w.WriteHeader(http.StatusCreated)
-}
-
-func handleDeleteRTCDService(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "delete-rtcd")
-	c.Ctx = logger.WithNamespace(c.Ctx, "rtcd")
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "rtcd")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	err = helmClient.UninstallReleaseByName("rtcd")
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to delete rtcd service")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func handleDeployRTCDService(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.Ctx = logger.WithField(c.Ctx, "action", "deploy-rtcd")
-	c.Ctx = logger.WithNamespace(c.Ctx, "mattermost-rtcd")
-
-	vars := mux.Vars(r)
-	clusterName := vars["name"]
-	if clusterName == "" || clusterName == "undefined" {
-		model.WriteBadRequestError(w, "Cluster name is required for RTCD deployment", "deploy-rtcd", map[string]interface{}{
-			"provided_cluster_name": clusterName,
-		})
-		return
-	}
-	c.Ctx = logger.WithClusterName(c.Ctx, clusterName)
-
-	helmClient, err := c.CloudProvider.HelmClient(c.Ctx, clusterName, "mattermost-rtcd")
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to authenticate helm client")
-		model.WriteDeploymentError(w, "Failed to connect to the Kubernetes cluster. Please verify your cluster is running and accessible.", "deploy-rtcd", map[string]interface{}{
-			"cluster_name": clusterName,
-			"namespace":    "mattermost-rtcd",
-			"error":        err.Error(),
-		})
-		return
-	}
-
-	chartRepo := repo.Entry{
-		Name: "mattermost",
-		URL:  "https://helm.mattermost.com",
-	}
-
-	err = helmClient.AddOrUpdateChartRepo(chartRepo)
-
-	if err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to add or update chart repo")
-		model.WriteDeploymentError(w, "Failed to add the Mattermost Helm repository. This may be due to network connectivity issues.", "deploy-rtcd", map[string]interface{}{
-			"repository_name": chartRepo.Name,
-			"repository_url":  chartRepo.URL,
-			"cluster_name":    clusterName,
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	// RTCD configuration based on Mattermost documentation
-	valuesYaml := helm.RTCDServiceValues
-
-	chartSpec := helmclient.ChartSpec{
-		ReleaseName:     "mattermost-rtcd",
-		ChartName:       "mattermost/mattermost-rtcd",
-		Namespace:       "mattermost-rtcd",
-		UpgradeCRDs:     true,
-		Wait:            true,
-		Timeout:         300 * time.Second,
-		CreateNamespace: true,
-		CleanupOnFail:   true,
-		ValuesYaml:      valuesYaml,
-	}
-
-	// Install a chart release.
-	// Note that helmclient.Options.Namespace should ideally match the namespace in chartSpec.Namespace.
-	if _, err := helmClient.InstallOrUpgradeChart(context.Background(), &chartSpec, nil); err != nil {
-		logger.FromContext(c.Ctx).WithError(err).Error("Failed to install rtcd service")
-		model.WriteDeploymentError(w, "Failed to deploy the RTCD service to your cluster. This could be due to insufficient resources, networking issues, or configuration problems.", "deploy-rtcd", map[string]interface{}{
-			"chart_name":      chartSpec.ChartName,
-			"release_name":    chartSpec.ReleaseName,
-			"namespace":       chartSpec.Namespace,
-			"cluster_name":    clusterName,
-			"timeout_seconds": int(chartSpec.Timeout.Seconds()),
-			"error":           err.Error(),
-		})
-		return
-	}
-
-	w.WriteHeader(http.StatusCreated)
 }
 
 func handleGetMattermostInstallations(c *Context, w http.ResponseWriter, r *http.Request) {
